04_use_cases/weather_agent_app/requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
chainlit>=0.7.700
pydantic>=2.5.0 
openai-agents>=0.1.0